# Conflicts:
#	cl_dll/hl/hl_baseentity.cpp
#	cl_dll/hl/hl_weapons.cpp
#	dlls/client.cpp
#	dlls/util.cpp
#	dlls/util.h
#	src/game/server/entities/NPCs/combat.cpp
#	src/game/server/entities/cbase.h
#	src/game/server/nodes.cpp

+++ src/game/server/client.cpp
@@ -47,10 +47,10 @@
 #include "ctf/CTFGoalFlag.h"
 #include "ctfplay_gamerules.h"
 
-extern DLL_GLOBAL ULONG		g_ulModelIndexPlayer;
+extern DLL_GLOBAL unsigned int		g_ulModelIndexPlayer;
 extern DLL_GLOBAL BOOL		g_fGameOver;
 extern DLL_GLOBAL int		g_iSkillLevel;
-extern DLL_GLOBAL ULONG		g_ulFrameCount;
+extern DLL_GLOBAL unsigned int		g_ulFrameCount;
 
 extern void CopyToBodyQue(entvars_t* pev);
 extern int giPrecacheGrunt;
